feat(characters): add character card component
